Modify: searchFilter 컴포넌트 SearchPage로 병합
