+++ src/pages/SearchPage/SearchPage.scss
@@ -71,6 +71,66 @@
   }
 }
 
+.filterContainer {
+  width: 100%;
+  height: 66px;
+  font-size: 13px;
+  display: flex;
+  align-items: center;
+  border-top: 1px solid #ececec;
+  .filterLists {
+    width: 1248px;
+    margin: 0 auto;
+    display: flex;
+    gap: 50px;
+    .filterList {
+      position: relative;
+      .filterListBtnOff {
+        width: 100px;
+        cursor: pointer;
+        text-align: start;
+        background-image: url('../../assets/image/toggleDown.png');
+        background-repeat: no-repeat;
+        background-size: 10px 8px;
+        background-position: 100% 50%;
+        background-color: transparent;
+      }
+
+      .filterListBtnOn {
+        width: 100px;
+        cursor: pointer;
+        text-align: start;
+        background-image: url('../../assets/image/toggleUp.png');
+        background-repeat: no-repeat;
+        background-size: 10px 8px;
+        background-position: 100% 50%;
+        background-color: transparent;
+      }
+
+      .filterListMenu {
+        width: 120px;
+        position: absolute;
+        border: 1px solid #dbdbdb;
+        background-color: #fff;
+        padding: 10px 10px 0 10px;
+        right: 0;
+        top: 30px;
+        & li {
+          margin-bottom: 10px;
+          cursor: pointer;
+        }
+        display: none;
+      }
+      .toggleListOn {
+        display: block;
+      }
+      .on {
+        font-weight: 700;
+      }
+    }
+  }
+}
+
 .noResult {
   width: 100%;
   height: 650px;
